backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
earthengine-api==0.1.384
pydantic==2.5.0
python-dotenv==1.0.0
